requirements.txt
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=9.0.0
python-dotenv>=1.0.0
anthropic>=0.25.0
openai>=1.0.0
